重构服务模块：将所有对geminiService的引用替换为aiService，简化服务管理并提升代码一致性，同时删除geminiService文件以清理项目结构。
